% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{accuracy}
\alias{accuracy}
\title{Accuracy (fraction of correctly classified obesravtions)}
\usage{
accuracy(y_true, y_pred)
}
\arguments{
\item{y_true}{boolean vector of target variable}

\item{y_pred}{boolean vector of predictions}
}
\description{
Accuracy (fraction of correctly classified obesravtions)
}
