% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sequences.R
\name{add_motifs}
\alias{add_motifs}
\title{injects motifs to a sequence
TODO: examples}
\usage{
add_motifs(motifs, sequence)
}
\arguments{
\item{motifs}{list of motifs to be injected}

\item{sequence}{vector of alphabet elements}
}
\value{
list(sequence, motifs, masks)
}
\description{
injects motifs to a sequence
TODO: examples
}
\examples{
# simple injection
add_motifs(list(c(1, "_", 1), c(1, 1)), c(2, 2, 3, 4))
# little bit more interesting
alph <- as.character(1L:4)
motifs <- generate_motifs(alph, 2)
example_sequence <- simulate_single_sequence(10, alph)
add_motifs(motifs, example_sequence)
}
