% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_analysis.R
\name{aggregate_results}
\alias{aggregate_results}
\title{aggregates results from `parse_results`}
\usage{
aggregate_results(df, ranking = TRUE)
}
\arguments{
\item{df}{output of `parse_results()`}
}
\value{
aggregated results (metrics means and SEs)
}
\description{
aggregates results from `parse_results`
}
