% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmark_summary}
\alias{benchmark_summary}
\title{Function summarizes results for a given feature selection method}
\usage{
benchmark_summary(scores, setup)
}
\description{
Function summarizes results for a given feature selection method
}
