% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_ranking.R
\name{build_model}
\alias{build_model}
\title{function builds a model and predicts out of fold probabilites}
\usage{
build_model(X_train, y_train, X_test, y_test, param, method)
}
\description{
function builds a model and predicts out of fold probabilites
}
