% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngram_probs.R
\name{calculate_ngram_prob}
\alias{calculate_ngram_prob}
\title{function computes probability of given n-gram}
\usage{
calculate_ngram_prob(mc, seq)
}
\arguments{
\item{mc}{markovchain object containing transition matrix}
}
\description{
function computes probability of given n-gram
}
\examples{
alphabet <- letters[1:4]
sequences <- matrix(sample(alphabet, size = 50, replace=TRUE), nrow = 5, ncol = 10)
mc <- sequenceTransitionMatrix(sequences, alphabet)
example_ngram <- sample(c(alphabet, "_"), size = 5, replace = TRUE)
print(example_ngram)
calculate_ngram_prob(mc, example_ngram)
}
