% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_utils.R
\name{calculate_score}
\alias{calculate_score}
\title{Calculates metrics for a given feature selection method}
\usage{
calculate_score(scores, setup)
}
\arguments{
\item{scores}{list of filter scores}

\item{setup}{evaluation details}
}
\description{
Function returns metrics average based on scores computed on n-gram matrices
}
