% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngram_probs.R
\name{calculate_seq_prob}
\alias{calculate_seq_prob}
\title{function computes probability of given sequence}
\usage{
calculate_seq_prob(mc, seq)
}
\arguments{
\item{mc}{markovchain object containing transition matrix}

\item{seq}{sequence of alphabet elements}
}
\description{
function computes probability of given sequence
}
\examples{
alphabet <- letters[1:4]
sequences <- matrix(sample(alphabet, size = 50, replace=TRUE), nrow = 5, ncol = 10)
mc <- sequenceTransitionMatrix(sequences, alphabet)
example_seq <- sample(alphabet, size = 5, replace = TRUE)
print(example_seq)
calculate_seq_prob(mc, example_seq)
}
