% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_analysis.R
\name{collect_filtering_times}
\alias{collect_filtering_times}
\title{function aggregates computation times}
\usage{
collect_filtering_times(paths)
}
\arguments{
\item{paths}{list of paths that contain filtering results}
}
\value{
computation times
}
\description{
function aggregates computation times
}
