% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{compute_metrics}
\alias{compute_metrics}
\title{Aggregated scores}
\usage{
compute_metrics(y_true, y_pred)
}
\arguments{
\item{y_true}{boolean vector of target variable}

\item{y_pred}{boolean vector of predictions}
}
\description{
Aggregated scores
}
