% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine_similarity.R
\name{cosine_similarity}
\alias{cosine_similarity}
\title{calculates cosine similarity of two vectors}
\usage{
cosine_similarity(x, y)
}
\arguments{
\item{x}{first vector}

\item{y}{second vector}
}
\value{
cosine similarity of two vectors
}
\description{
calculates cosine similarity of two vectors
}
\examples{
cosine_similarity(1:4, 5:8)
}
