% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_simulation_data_setOfMotifs.R
\name{create_simulation_data_set_of_motifs}
\alias{create_simulation_data_set_of_motifs}
\title{function generates set of RDS files containing n-gram matrices}
\usage{
create_simulation_data_set_of_motifs(
  replications,
  seq_nums,
  seq_lengths,
  motif_set_size,
  num_motifs,
  alphabet,
  path,
  title,
  save_files = FALSE,
  motifProbs = NULL,
  seqProbs = NULL,
  fraction = 0.5,
  n = 4,
  d = 6
)
}
\arguments{
\item{replications}{number of matrices' generation for each set of parameters}

\item{seq_nums}{vector of sequences' numbers}

\item{seq_lengths}{vector of sequences' lengths}

\item{motif_set_size}{number of motifs to be created}

\item{num_motifs}{vector of number of motifs to be injected}

\item{alphabet}{elements used to build both sequence and motif}

\item{path}{file path of RDS files}

\item{title}{simulation name}

\item{save_files}{triggers if RDS will be saved}

\item{motifProbs}{alphabet probabilites for motifs}

\item{seqProbs}{alphabet probabilites for sequences}

\item{fraction}{fraction of positive sequences}

\item{n}{maximum number of alphabet elements in motif}

\item{d}{maximum number of gaps in motif}
}
\value{
data frame with RDS files' details
}
\description{
function generates set of RDS files containing n-gram matrices
}
