% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine_similarity.R
\name{euclidean_norm}
\alias{euclidean_norm}
\title{calculates euclidean norm of a given vector}
\usage{
euclidean_norm(x)
}
\arguments{
\item{x}{vector}
}
\value{
euclidean norm of `x`
}
\description{
calculates euclidean norm of a given vector
}
\examples{
euclidean_norm(1:4)
}
