% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_ranking.R
\name{evaluate_selected_kmers}
\alias{evaluate_selected_kmers}
\title{Evaluation of filtered k-mers in ranking model approach}
\usage{
evaluate_selected_kmers(df, validation_scheme)
}
\arguments{
\item{df}{data frame containing both selected k-mers and target variable y}

\item{validation_scheme}{list containing folds, n_kmers, cv_reps - validation setup}
}
\description{
Evaluation of filtered k-mers in ranking model approach
}
