% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{F1score}
\alias{F1score}
\title{F1-score (harmonic mean of precision and recall)}
\usage{
F1score(y_true, y_pred)
}
\arguments{
\item{y_true}{boolean vector of target variable}

\item{y_pred}{boolean vector of predictions}
}
\description{
F1-score (harmonic mean of precision and recall)
}
