% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_utils.R
\name{filter_ngrams}
\alias{filter_ngrams}
\title{Wrapper for variuous feature selection method evaluated in a benchmark}
\usage{
filter_ngrams(ngram_matrix, feature_selection_method)
}
\arguments{
\item{ngram_matrix}{matrix of n-gram occurences}

\item{feature_selection_method}{feature selection method name (QuiPT, ...)}
}
\description{
Wrapper for variuous feature selection method evaluated in a benchmark
}
