% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_ranking.R
\name{filter_nonrankings}
\alias{filter_nonrankings}
\title{function that evaluates nonranking methods}
\usage{
filter_nonrankings(
  paths,
  feature_selection_method,
  n,
  fraction,
  validation_scheme,
  thresholds = NULL
)
}
\arguments{
\item{paths}{list of paths containing n-gram matrices}

\item{feature_selection_method}{filter (e.g. QuiPT)}

\item{n}{number of total sequences}

\item{fraction}{fraction of positive examples}

\item{validation_scheme}{list with ranking details}

\item{thresholds}{p-value thresholds for statistical tests}
}
\description{
function that evaluates nonranking methods
}
