% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_ranking.R
\name{filter_nonrankings_exp3}
\alias{filter_nonrankings_exp3}
\title{function that evaluates nonranking methods}
\usage{
filter_nonrankings_exp3(
  paths,
  num_reps,
  num_matrices_to_rbind,
  output_prefix,
  feature_selection_method,
  n,
  fraction,
  validation_scheme,
  thresholds = NULL
)
}
\arguments{
\item{paths}{list of paths containing n-gram matrices}

\item{num_reps}{number of repetitions of an experiment}

\item{num_matrices_to_rbind}{number of matrices to bind in order to perform exp III correctly}

\item{output_prefix}{output files directory}

\item{feature_selection_method}{filter (e.g. QuiPT)}

\item{n}{number of total sequences}

\item{fraction}{fraction of positive examples}

\item{validation_scheme}{list with ranking details}

\item{thresholds}{p-value thresholds for statistical tests}
}
\description{
function that evaluates nonranking methods
}
