% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_ranking.R
\name{filter_rankings_exp3}
\alias{filter_rankings_exp3}
\title{function creates and evaluates filtering rankings}
\usage{
filter_rankings_exp3(
  paths,
  num_reps,
  num_matrices_to_rbind,
  output_prefix,
  feature_selection_method,
  n,
  fraction,
  validation_scheme
)
}
\arguments{
\item{paths}{list of paths containing n-gram matrices}

\item{num_reps}{number of repetitions of an experiment}

\item{num_matrices_to_rbind}{number of matrices to bind in order to perform exp III correctly}

\item{output_prefix}{output files directory}

\item{feature_selection_method}{filter (e.g. QuiPT)}

\item{n}{number of total sequences}

\item{fraction}{fraction of positive examples}

\item{validation_scheme}{list with ranking details}
}
\description{
function creates and evaluates filtering rankings
}
