% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sequences.R
\name{generate_motifs}
\alias{generate_motifs}
\title{generate multiple motifs from alphabet}
\usage{
generate_motifs(alphabet, n_motif, n = 4, d = 6, motifProbs = NULL)
}
\arguments{
\item{alphabet}{elements used to build motif}

\item{n_motif}{number of motifs to be generated}

\item{n}{maximum number of alphabet elements in motif}

\item{d}{maximum number of gaps in motif}

\item{motifProbs}{alphabet probabilites for motifs}
}
\value{
list of generated motifs
}
\description{
generate multiple motifs from alphabet
}
\examples{
generate_motifs(1:4, 5)
generate_motifs(1:4, 5, n = 6, d = 2, motifProbs = c(0.7, 0.1, 0.1, 0.1))
}
