% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine_similarity.R
\name{generate_probs}
\alias{generate_probs}
\title{function generates two vectors of probabilities with given cosine similarity}
\usage{
generate_probs(size, cosine_sim)
}
\arguments{
\item{size}{expected size of probabilities' vectors}

\item{cosine_sim}{cosine similarity of two vectors}
}
\value{
matrix with probabilites in rows
}
\description{
function generates two vectors of probabilities with given cosine similarity
}
\examples{
generate_probs(5, 0.9)
}
