% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sequences.R
\name{generate_sequences}
\alias{generate_sequences}
\title{function counts n-grams in given sequences}
\usage{
generate_sequences(
  n_seq,
  l_seq,
  alphabet,
  motifs_list,
  n_motifs,
  fraction = 0.5,
  seqProbs = NULL,
  n = 4,
  d = 6
)
}
\arguments{
\item{n_seq}{number of sequences to be generated}

\item{l_seq}{sequence length}

\item{alphabet}{elements used to build sequence}

\item{motifs_list}{list of injected motifs}

\item{n_motifs}{number of motifs injected to each positive sequence}

\item{fraction}{TODO: add fraction: of positive sequences / change approach}

\item{seqProbs}{alphabet probabilites for sequences}

\item{n}{maximum number of alphabet elements in n-gram}

\item{d}{maximum number of gaps in n-gram}
}
\value{
generated sequences
}
\description{
function counts n-grams in given sequences
}
\examples{
n_seq <- 20
len <- 1200
alph <- letters[1:4]
motifs <- generate_motifs(alph, 2)
results <- generate_sequences(n_seq, len, alph, motifs, 1)
results <- generate_sequences(n_seq, len, alph, motifs, 1, seqProbs = c(0.7, 0.1, 0.1, 0.1))
results
attributes(results)

}
