% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_ranking.R
\name{kmers_for_nonranking_methods}
\alias{kmers_for_nonranking_methods}
\title{calculates number of k-mers to select for non-ranking methods}
\usage{
kmers_for_nonranking_methods(filtering_results, method, thresholds)
}
\arguments{
\item{filtering_results}{data.frame containing ranking of k-mers}

\item{method}{feature filtering method}
}
\description{
calculates number of k-mers to select for non-ranking methods
}
