% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_utils.R
\name{positive_ngrams}
\alias{positive_ngrams}
\title{Function returns ector indicating n-grams considered positive}
\usage{
positive_ngrams(ngram_matrix)
}
\description{
Positive n-gram is either n-gram containing motif (not longer than n + (n+1) // 4)
or n-gram that is a part of motif
}
