% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{precision}
\alias{precision}
\title{Precision}
\usage{
precision(y_true, y_pred)
}
\arguments{
\item{y_true}{boolean vector of target variable}

\item{y_pred}{boolean vector of predictions}
}
\description{
Precision
}
