% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuiPTevaluation.R
\name{QuiPT_summary}
\alias{QuiPT_summary}
\title{test QuiPT}
\usage{
QuiPT_summary(ngram_matrix)
}
\arguments{
\item{ngram_matrix}{matrix of n-gram occurences (computed in [create_simulation_data()])}

\item{mc}{markovchain object to calculate n-gram probability}

\item{adjustments}{p-value adjustments for multiple testing}

\item{thresholds}{p-value thresholds}
}
\value{
data frame with results of QuiPT testing
}
\description{
test QuiPT
}
