% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{QuiPTsimBenchmark}
\alias{QuiPTsimBenchmark}
\title{Function wraps up QuiPT evaluation auxiliary functions}
\usage{
QuiPTsimBenchmark(paths, setup)
}
\arguments{
\item{paths}{list of paths of n-gram matrices}

\item{setup}{list of experiment details}
}
\description{
Function wraps up QuiPT evaluation auxiliary functions
}
\seealso{
`create_benchmark_data` `benchmark_summary`
}
