% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_ranking.R
\name{ranking_summary}
\alias{ranking_summary}
\title{High-level wrapper of ranking evaluation tools}
\usage{
ranking_summary(paths, setup, validation_scheme)
}
\arguments{
\item{setup}{filter usage setup}

\item{validation_scheme}{list containing folds, n_kmers, cv_reps - validation setup}
}
\description{
High-level wrapper of ranking evaluation tools
}
