% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuiPTevaluation.R
\name{rbind_ngram_matrices}
\alias{rbind_ngram_matrices}
\title{function binds n-gram matrices}
\usage{
rbind_ngram_matrices(...)
}
\arguments{
\item{...}{ngram_matrices (list of matrices can also be passed)}
}
\value{
combined ngram matrix
}
\description{
function binds n-gram matrices
}
