% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuiPTevaluation.R
\name{rbind_ngrams}
\alias{rbind_ngrams}
\title{function combines two ngram matrices}
\usage{
rbind_ngrams(m1, m2)
}
\arguments{
\item{m1}{upper matrix}

\item{m2}{lower matrix}
}
\value{
combined ngram matrix
}
\description{
function combines two ngram matrices
}
