% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_ranking.R
\name{rbind_results}
\alias{rbind_results}
\title{functions binds rows of data frames in a smart way}
\usage{
rbind_results(...)
}
\arguments{
\item{...}{dataframes to bind}
}
\description{
functions binds rows of data frames in a smart way
}
