% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_ranking.R
\name{rbind_smart}
\alias{rbind_smart}
\title{functions binds rows of two data frames in a smart way
(non overlapping columns are filled with NAs)}
\usage{
rbind_smart(df1, df2)
}
\description{
functions binds rows of two data frames in a smart way
(non overlapping columns are filled with NAs)
}
