% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuiPTevaluation.R
\name{read_ngram_matrix}
\alias{read_ngram_matrix}
\title{function reads output files of `create_simulation_data`}
\usage{
read_ngram_matrix(filePath, n = NULL, fraction = 0.5)
}
\arguments{
\item{filePath}{path of RDS file}

\item{n}{number of sequences to be subsetted (optional)}

\item{fraction}{fraction of positive sequences (optional)}
}
\description{
function reads output files of `create_simulation_data`
}
\seealso{
[create_simulation_data()] [subset_matrix()]
}
