% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuiPTevaluation.R
\name{read_simulation_data}
\alias{read_simulation_data}
\title{function reads results of `create_simulation_data` function
attribute contains simulation details}
\usage{
read_simulation_data(path, title)
}
\arguments{
\item{path}{directory containing result files}

\item{title}{simulation title}
}
\description{
function reads results of `create_simulation_data` function
attribute contains simulation details
}
