% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{sensitivity}
\alias{sensitivity}
\title{Sensitivity (true positive rate, recall)}
\usage{
sensitivity(y_true, y_pred)
}
\arguments{
\item{y_true}{boolean vector of target variable}

\item{y_pred}{boolean vector of predictions}
}
\description{
Sensitivity (true positive rate, recall)
}
