% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngram_probs.R
\name{sequenceTransitionMatrix}
\alias{sequenceTransitionMatrix}
\title{function computes transition matrix}
\usage{
sequenceTransitionMatrix(sequences, alphabet)
}
\arguments{
\item{sequences}{matrix of sequences (each row represents single sequence)}

\item{alphabet}{elements used to build sequences}
}
\value{
combined ngram matrix
}
\description{
function computes transition matrix
}
\examples{
alphabet <- letters[1:4]
sequences <- matrix(sample(alphabet, size = 50, replace=TRUE), nrow = 5, ncol = 10)
sequenceTransitionMatrix(sequences, alphabet)

}
