% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sequences.R
\name{simulate_sequences}
\alias{simulate_sequences}
\title{function generates sequences (both positive & negative)}
\usage{
simulate_sequences(
  n_seq,
  len,
  alphabet,
  motifs_list,
  n_motifs,
  fraction = 0.5,
  seqProbs = NULL
)
}
\arguments{
\item{n_seq}{number of sequences to be generated}

\item{len}{sequence length}

\item{alphabet}{elements used to build sequence}

\item{motifs_list}{list of injected motifs}

\item{n_motifs}{number of motifs injected to each positive sequence}

\item{fraction}{of positive sequences}

\item{seqProbs}{alphabet probabilites for sequences}
}
\value{
generated sequences
}
\description{
function generates sequences (both positive & negative)
}
\examples{
n_seq <- 20
len <- 10
alph <- 1L:4
motifs <- generate_motifs(alph, 3)
simulate_sequences(n_seq, len, alph, motifs, 1)
simulate_sequences(n_seq, len, alph, motifs, 1, fraction = 0.8)
simulate_sequences(n_seq, len, alph, motifs, 1, seqProbs = c(0.7, 0.1, 0.1, 0.1))
}
