% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sequences.R
\name{simulate_single_sequence}
\alias{simulate_single_sequence}
\title{generates sequence of elements from alphabet with replacement}
\usage{
simulate_single_sequence(len, alphabet)
}
\arguments{
\item{len}{length of generated sample sequence}

\item{alphabet}{elements used to build sequence}
}
\value{
randomly generated sequence
}
\description{
generates sequence of elements from alphabet with replacement
}
\examples{
simulate_single_sequence(5, 1L:4)
simulate_single_sequence(10, c("a", "b", "c"))
}
