% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuiPTevaluation.R
\name{subset_matrix}
\alias{subset_matrix}
\title{subsets simple triplet matrix}
\usage{
subset_matrix(filePath, n, fraction)
}
\arguments{
\item{filePath}{path of RDS file}

\item{n}{number of sequences to be subsetted}

\item{fraction}{fraction of positive sequences}
}
\description{
subsets simple triplet matrix
}
