% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_ranking.R
\name{build_model}
\alias{build_model}
\title{function builds a model and predicts out of fold probabilites}
\usage{
build_model(X_train, y_train, X_test, y_test, p, method)
}
\arguments{
\item{p}{vector of parameters (lm : lambda, knn: neighbors, naive bayes: laplace, rf: num.trees)}

\item{method}{model (one of: lm, knn, naive bayes, rf)}
}
\description{
function builds a model and predicts out of fold probabilites
}
