% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{create_benchmark_data}
\alias{create_benchmark_data}
\title{Computes feature selection on n-gram matrices
Result is an input to the benchmark.}
\usage{
create_benchmark_data(paths, setup)
}
\description{
Computes feature selection on n-gram matrices
Result is an input to the benchmark.
}
