% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_simulation_data.R
\name{create_simulation_data}
\alias{create_simulation_data}
\title{function generates set of RDS files containing n-gram matrices}
\usage{
create_simulation_data(
  replications,
  seq_nums,
  seq_lengths,
  num_motifs,
  alphabet,
  path,
  title,
  save_files = FALSE,
  motifProbs = NULL,
  seqProbs = NULL,
  fraction = 0.5,
  n = 4,
  d = 6
)
}
\arguments{
\item{replications}{number of matrices' generation for each set of parameters}

\item{seq_nums}{vector of sequences' numbers}

\item{seq_lengths}{vector of sequences' lengths}

\item{num_motifs}{vector of number of motifs to be injected}

\item{alphabet}{elements used to build both sequence and motif}

\item{path}{file path of RDS files}

\item{title}{simulation name}

\item{save_files}{triggers if RDS will be saved}

\item{motifProbs}{alphabet probabilites for motifs}

\item{seqProbs}{alphabet probabilites for sequences}

\item{fraction}{fraction of positive sequences}

\item{n}{maximum number of alphabet elements in motif}

\item{d}{maximum number of gaps in motif}
}
\value{
data frame with RDS files' details
}
\description{
function generates set of RDS files containing n-gram matrices
}
\examples{
alph <- 1L:4
reps <- 10
n_seq <- c(10)
l_seq <- c(10)
n_motifs <- c(1, 2)
path = "./"
results <- create_simulation_data(reps, n_seq, l_seq, n_motifs, alph, path, "SEQ",FALSE)

alph <- 1L:4
reps <- 3
n_seq <- c(10)
l_seq <- c(10)
n_motifs <- 1
results <- create_simulation_data(reps, n_seq, l_seq, n_motifs, alph,
                                 path, "SEQ", FALSE,
                                 motifProbs = c(0.7, 0.1, 0.1, 0.1),
                                 seqProbs = c(0.7, 0.1, 0.1, 0.1),
                                 n = 4, d = 4)

}
