% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_ranking.R
\name{evaluate_models}
\alias{evaluate_models}
\title{Function trains and evaluates model on selected k-mers}
\usage{
evaluate_models(df_train, df_test, validation_scheme)
}
\description{
Function trains and evaluates model on selected k-mers
}
