% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sequences.R
\name{generate_motifs}
\alias{generate_motifs}
\title{generate multiple motifs from alphabet}
\usage{
generate_motifs(alphabet, n_motif)
}
\arguments{
\item{alphabet}{elements used to build motif}

\item{n_motif}{number of motifs to be generated}
}
\value{
list of generated motifs
}
\description{
generate multiple motifs from alphabet
}
\examples{
generate_motifs(1:4, 5)
}
