% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sequences.R
\name{generate_sequences}
\alias{generate_sequences}
\title{function counts n-grams in given sequences}
\usage{
generate_sequences(n_seq, l_seq, alphabet, motifs_list, n_motifs)
}
\arguments{
\item{n_seq}{number of sequences to be generated}

\item{l_seq}{sequence length}

\item{alphabet}{elements used to build sequence}

\item{motifs_list}{list of injected motifs}

\item{n_motifs}{number of motifs injected to each positive sequence}

\item{fraction}{TODO: add fraction: of positive sequences / change approach}
}
\value{
generated sequences
}
\description{
function counts n-grams in given sequences
}
\examples{
n_seq <- 20
len <- 10
alph <- 1L:4
motifs <- generate_motifs(alph, 3)
generate_sequences(n_seq, len, alph, motifs, 1)
}
