% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sequences.R
\name{generate_single_motif}
\alias{generate_single_motif}
\title{function generates motif from an alphabet
TODO: motif length control}
\usage{
generate_single_motif(alphabet, n = 4, d = 6, motifProbs = NULL)
}
\arguments{
\item{alphabet}{elements used to build motif}

\item{n}{number of elements from alphabet}

\item{d}{total sum of gaps}
}
\value{
motif based on given alphabet
}
\description{
function generates motif from an alphabet
TODO: motif length control
}
\examples{
generate_single_motif(1:4)
}
