% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sequences.R
\name{generate_single_motif}
\alias{generate_single_motif}
\title{function generates motif from an alphabet
TODO: motif length control}
\usage{
generate_single_motif(alphabet)
}
\arguments{
\item{alphabet}{elements used to build motif}
}
\value{
motif based on given alphabet
}
\description{
function generates motif from an alphabet
TODO: motif length control
}
\examples{
generate_single_motif(1:4)
}
